api/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
flake8==7.0.0
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
